call function directly
